// client/src/components/location/__generated__/LocationList_query.graphql.ts
export type LocationList_query$data = {
  readonly clogLocationSearch: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly createdAt: string;
        readonly id: string;
        readonly name: string;
        readonly stockCount: number;
      } | null | undefined;
    } | null | undefined> | null | undefined;
    readonly totalCount: number | null | undefined;
  } | null | undefined;
  readonly " $fragmentType": "LocationList_query";
};
export type LocationList_query$key = {
  readonly " $data"?: LocationList_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"LocationList_query">;
};